requirements.txt
PySide6==6.6.1
keyboard==0.13.5
pillow==10.2.0
pywin32==306
openai==1.12.0
python-dotenv==1.0.0
numpy==1.26.3
python-tesseract
opencv-python
numpy
pillow
pynput